chore: fix postgres update issues

+++ builder/img/busybox.sh
@@ -59,7 +59,9 @@ cp /lib/${ARCH_LIB_DIR}/${LOADER} lib
 if ! mountpoint -q /var/cache/apt/archives 2>/dev/null; then
   rm -rf /var/cache/apt/archives/* "/var/cache/apt/archives/partial"/*
 fi
-rm -rf /var/lib/apt/lists/*
+if ! mountpoint -q /var/lib/apt/lists 2>/dev/null; then
+  rm -rf /var/lib/apt/lists/*
+fi
 
 # Build squashfs
 mksquashfs "${TMP}/root" "/mnt/out/layer.squashfs" -noappend
