+++ builder/img/go.sh
@@ -32,4 +32,6 @@ git reset --hard ${gobin_commit}
 if ! mountpoint -q /var/cache/apt/archives 2>/dev/null; then
   rm -rf /var/cache/apt/archives/* "/var/cache/apt/archives/partial"/*
 fi
-rm -rf /var/lib/apt/lists/*
+if ! mountpoint -q /var/lib/apt/lists 2>/dev/null; then
+  rm -rf /var/lib/apt/lists/*
+fi
