+++ builder/img/protoc.sh
@@ -44,4 +44,6 @@ GOFLAGS=-mod=mod GOBIN=/bin go install google.golang.org/grpc/cmd/protoc-gen-go-
 if ! mountpoint -q /var/cache/apt/archives 2>/dev/null; then
   rm -rf /var/cache/apt/archives/* "/var/cache/apt/archives/partial"/*
 fi
-rm -rf /var/lib/apt/lists/*
+if ! mountpoint -q /var/lib/apt/lists 2>/dev/null; then
+  rm -rf /var/lib/apt/lists/*
+fi
