+++ test/img/packages.sh
@@ -23,4 +23,6 @@ git config --global "user.name"  "Flynn Test"
 if ! mountpoint -q /var/cache/apt/archives 2>/dev/null; then
   rm -rf /var/cache/apt/archives/* "/var/cache/apt/archives/partial"/*
 fi
-rm -rf /var/lib/apt/lists/*
+if ! mountpoint -q /var/lib/apt/lists 2>/dev/null; then
+  rm -rf /var/lib/apt/lists/*
+fi
